# Conflicts:
#	src/Umbraco.Core/DependencyInjection/UmbracoBuilder.cs

+++ src/Umbraco.Core/Services/WebhookService.cs
@@ -1,25 +1,42 @@
+using Umbraco.Cms.Core.Events;
 using Umbraco.Cms.Core.Models;
+using Umbraco.Cms.Core.Notifications;
 using Umbraco.Cms.Core.Persistence.Repositories;
 using Umbraco.Cms.Core.Scoping;
 
 namespace Umbraco.Cms.Core.Services;
 
-public class WebhookService : IWebHookService
+public class WebhookService : IWebhookService
 {
     private readonly ICoreScopeProvider _provider;
     private readonly IWebhookRepository _webhookRepository;
+    private readonly IEventMessagesFactory _eventMessagesFactory;
 
-    public WebhookService(ICoreScopeProvider provider, IWebhookRepository webhookRepository)
+    public WebhookService(ICoreScopeProvider provider, IWebhookRepository webhookRepository, IEventMessagesFactory eventMessagesFactory)
     {
         _provider = provider;
         _webhookRepository = webhookRepository;
+        _eventMessagesFactory = eventMessagesFactory;
     }
 
     /// <inheritdoc />
-    public async Task<Webhook> CreateAsync(Webhook webhook)
+    public async Task<Webhook?> CreateAsync(Webhook webhook)
     {
         using ICoreScope scope = _provider.CreateCoreScope();
+
+        EventMessages eventMessages = _eventMessagesFactory.Get();
+        var savingNotification = new WebhookSavingNotification(webhook, eventMessages);
+        if (scope.Notifications.PublishCancelable(savingNotification))
+        {
+            scope.Complete();
+            return null;
+        }
+
         Webhook created = await _webhookRepository.CreateAsync(webhook);
+
+        scope.Notifications.Publish(
+            new WebhookSavedNotification(webhook, eventMessages).WithStateFrom(savingNotification));
+
         scope.Complete();
 
         return created;
@@ -37,6 +54,14 @@ public class WebhookService : IWebHookService
             throw new ArgumentException("Webhook does not exist");
         }
 
+        EventMessages eventMessages = _eventMessagesFactory.Get();
+        var savingNotification = new WebhookSavingNotification(webhook, eventMessages);
+        if (scope.Notifications.PublishCancelable(savingNotification))
+        {
+            scope.Complete();
+            return;
+        }
+
         currentWebhook.Enabled = webhook.Enabled;
         currentWebhook.ContentTypeKeys = webhook.ContentTypeKeys;
         currentWebhook.Events = webhook.Events;
@@ -44,6 +69,10 @@ public class WebhookService : IWebHookService
         currentWebhook.Headers = webhook.Headers;
 
         await _webhookRepository.UpdateAsync(currentWebhook);
+
+        scope.Notifications.Publish(
+            new WebhookSavedNotification(webhook, eventMessages).WithStateFrom(savingNotification));
+
         scope.Complete();
     }
 
@@ -54,7 +83,17 @@ public class WebhookService : IWebHookService
         Webhook? webhook = await _webhookRepository.GetAsync(key);
         if (webhook is not null)
         {
+            EventMessages eventMessages = _eventMessagesFactory.Get();
+            var deletingNotification = new WebhookDeletingNotification(webhook, eventMessages);
+            if (scope.Notifications.PublishCancelable(deletingNotification))
+            {
+                scope.Complete();
+                return;
+            }
+
             await _webhookRepository.DeleteAsync(webhook);
+            scope.Notifications.Publish(
+                new WebhookDeletedNotification(webhook, eventMessages).WithStateFrom(deletingNotification));
         }
 
         scope.Complete();
